Rename popover filename
